feat/blog_image upload

+++ services/blog_imageService.js
@@ -1,14 +1,51 @@
 import supabase from "../config/supabaseClient.js";
+import { v4 as uuidv4 } from "uuid";
 
-export const createBlog_image = async (blog_imageData) => {
+
+export const uploadBlogImage = async (blogImageData, file) => {
+  try{
+   const fileExt = file.originalname.split(".").pop();
+   const fileName = `${blogImageData.blog_id}-${uuidv4()}.${fileExt}`;
+   const filePath = `blog/${fileName}`;
+
+  const { error: uploadError } = await supabase.storage
+    .from("images")
+    .upload(filePath, file.buffer, {
+      cacheControl: "3600",
+      upsert: false,
+      contentType: file.mimetype,
+    });
+
+  if (uploadError) throw new Error(uploadError.message);
+  const { data: publicUrlData } = supabase.storage
+    .from("images")
+    .getPublicUrl(filePath);
+
+  const url = publicUrlData.publicUrl;
+
+  // Insert into DB
   const { data, error } = await supabase
     .from("blog_images")
-    .insert([blog_imageData])
-    .select();
+    .insert([
+      {
+        url: url,
+        blog_id: blogImageData.blog_id,
+        filename: fileName,
+      },
+    ])
+    .select()
+    .single();
 
   if (error) throw new Error(error.message);
-  return data;
+  return data
+  } catch (err) {
+    throw new Error(err.message);
+  }
 };
+
+
+
+
 
 export const getBlog_image = async () => {
   const { data, error } = await supabase.from("blog_images").select("*");
@@ -22,7 +59,7 @@ export const getBlog_imageById = async (id) => {
     .from("blog_images")
     .select("*")
     .eq("id", id)
-    .single();
+    .maybeSingle();
 
   if (error) throw new Error(error.message);
   return data;
@@ -35,20 +72,30 @@ export const updateBlog_image = async (id, updates) => {
     .from("blog_images")
     .update(updates)
     .eq("id", id)
-    .select();
+    .select()
+    .single();
 
   if (error) throw new Error(error.message);
   return data;
 };
 
 export const deleteBlog_image = async (id) => {
+  const existing = await getBlog_imageById(id);
+  if (!existing) throw new Error("Blog image not found");
+  const { error: storageError } = await supabase.storage
+    .from("images")
+    .remove([`blog/${existing.filename}`]);
+
+  if (storageError) throw new Error(storageError.message);
+
+
   const { error } = await supabase
     .from("blog_images")
     .delete()
     .eq("id", id)
-    .select();
 
   if (error) throw new Error(error.message);
   return { success: true, message: "blogs image deleted successfully" };
 };
+
